+++ code-gen/update-cluster-state.sh
@@ -5,17 +5,17 @@
 
 # This script may be used to upgrade an existing cluster state repo. It is designed to be non-destructive in that it
 # won't push any changes to the server. Instead, it will set up a parallel branch for every CDE branch and/or the
-# customer-hub branch as specified through the ENVIRONMENTS environment variable. For example, if the new version is
-# v1.7.1 and the ENVIRONMENTS variable override is not provided, then it’ll set up 4 new CDE branches at the new
-# version for the default set of environments: v1.7.1-dev, v1.7.1-test, v1.7.1-stage and v1.7.1-master and 1 new
-# customer-hub branch v1.7.1-customer-hub.
+# customer-hub branch as specified through the SUPPORTED_ENVIRONMENT_TYPES environment variable. For example, if the
+# new version is v1.7.1 and the SUPPORTED_ENVIRONMENT_TYPES variable override is not provided, then it’ll set up 4 new
+# CDE branches at the new version for the default set of environments: v1.7.1-dev, v1.7.1-test, v1.7.1-stage and
+# v1.7.1-master and 1 new customer-hub branch v1.7.1-customer-hub.
 
 # NOTE: The script must be run from the root of the cluster state repo clone directory. It acts on the following
 # environment variables.
 #
 #   NEW_VERSION -> Required. The new version of Beluga to which to update the cluster state repo.
-#   ENVIRONMENTS -> A space-separated list of environments. Defaults to 'dev test stage prod customer-hub', if unset.
-#       If provided, it must contain all or a subset of the environments currently created by the
+#   SUPPORTED_ENVIRONMENT_TYPES -> A space-separated list of environments. Defaults to 'dev test stage prod customer-hub',
+#       if unset. If provided, it must contain all or a subset of the environments currently created by the
 #       generate-cluster-state.sh script, i.e. dev, test, stage, prod and customer-hub.
 #   RESET_TO_DEFAULT -> An optional flag, which if set to true will reset the cluster-state-repo to the OOTB state
 #       for the new version. This has the same effect as running the platform code build job that initially seeds the
@@ -69,8 +69,15 @@ RESET_TO_DEFAULT="${RESET_TO_DEFAULT:-false}"
 #     find "${K8S_CONFIGS_DIR}" -type f -exec basename {} + | sort -u   # Run this command on each tag
 #     cat v1.7-k8s-files v1.8-k8s-files | sort -u                       # Create a union of the k8s files
 
+# TODO: `argo-application.yaml` included here to prevent unintentionally adding it to custom_resources when upgrading
+# from pre-1.18 versions. Remove it when all customers are on 1.18.
+# Jira ticket: https://pingidentity.atlassian.net/browse/PDO-5247
+
 beluga_owned_k8s_files="@.flux.yaml \
 @argo-application.yaml \
+@argocd-application-set.yaml \
+@argocd-cm-patch.yaml \
+@argocd-strategic-patches.yaml \
 @custom-patches-sample.yaml \
 @env_vars \
 @flux-command.sh \
@@ -517,9 +524,6 @@ handle_changed_k8s_configs() {
   log "DEBUG: Found the following new files in branch '${OLD_BRANCH}':"
   echo "${new_files}"
 
-  KUSTOMIZATION_FILE="${CUSTOM_RESOURCES_REL_DIR}/kustomization.yaml"
-  KUSTOMIZATION_BAK_FILE="${KUSTOMIZATION_FILE}.bak"
-
   # Special-case the handling all files owned by PS/GSO.
 
   # 1. Copy the custom-patches.yaml file (owned by PS/GSO) as is.
@@ -559,19 +563,6 @@ handle_changed_k8s_configs() {
       mkdir -p "${new_file_dirname}"
       git show "${OLD_BRANCH}:${new_file}" > "${new_file}"
       continue
-    fi
-
-    log "Copying custom file ${OLD_BRANCH}:${new_file} into directory ${CUSTOM_RESOURCES_REL_DIR}"
-    git show "${OLD_BRANCH}:${new_file}" > "${CUSTOM_RESOURCES_REL_DIR}/${new_file_basename}"
-
-    log "Adding new resource file ${new_file_basename} to ${KUSTOMIZATION_FILE}"
-    new_resource_line="- ${new_file_basename}"
-
-    grep_opts=(-q -e "${new_resource_line}")
-    if ! grep "${grep_opts[@]}" "${KUSTOMIZATION_FILE}"; then
-      # shellcheck disable=SC1003
-      sed -i.bak -e '/^resources:$/a\'$'\n'"${new_resource_line}" "${KUSTOMIZATION_FILE}"
-      rm -f "${KUSTOMIZATION_BAK_FILE}"
     fi
   done
 
@@ -779,12 +770,12 @@ CURRENT_BRANCH="$(git rev-parse --abbrev-ref HEAD)"
 
 # Validate that a git branch exists for every environment.
 ALL_ENVIRONMENTS='dev test stage prod customer-hub'
-ENVIRONMENTS="${ENVIRONMENTS:-${ALL_ENVIRONMENTS}}"
+SUPPORTED_ENVIRONMENT_TYPES="${SUPPORTED_ENVIRONMENT_TYPES:-${ALL_ENVIRONMENTS}}"
 
 NEW_BRANCHES=
 REPO_STATUS=0
 
-for ENV in ${ENVIRONMENTS}; do
+for ENV in ${SUPPORTED_ENVIRONMENT_TYPES}; do
   test "${ENV}" = 'prod' &&
       OLD_BRANCH='master' ||
       OLD_BRANCH="${ENV}"
@@ -844,7 +835,7 @@ fi
 # For each environment:
 #   - Generate code for all its regions
 #   - Push code for all its regions into new branches
-for ENV in ${ENVIRONMENTS}; do # ENV loop
+for ENV in ${SUPPORTED_ENVIRONMENT_TYPES}; do # ENV loop
   test "${ENV}" = 'prod' &&
       OLD_BRANCH='master' ||
       OLD_BRANCH="${ENV}"
@@ -930,7 +921,7 @@ for ENV in ${ENVIRONMENTS}; do # ENV loop
             SERVER_PROFILE_URL='' \
             K8S_GIT_URL="${PING_CLOUD_BASE_REPO_URL}" \
             K8S_GIT_BRANCH="${NEW_VERSION}" \
-            ENVIRONMENTS="${NEW_BRANCH}" \
+            SUPPORTED_ENVIRONMENT_TYPES="${NEW_BRANCH}" \
             PING_IDENTITY_DEVOPS_USER='' \
             PING_IDENTITY_DEVOPS_KEY='' \
             MYSQL_USER='' \
@@ -1057,7 +1048,7 @@ for ENV in ${ENVIRONMENTS}; do # ENV loop
       QUIET=true \
           GENERATED_CODE_DIR="${TARGET_DIR}" \
           IS_PRIMARY=${IS_PRIMARY} \
-          ENVIRONMENTS="${NEW_BRANCH}" \
+          SUPPORTED_ENVIRONMENT_TYPES="${NEW_BRANCH}" \
           PUSH_TO_SERVER=false \
           "${NEW_PING_CLOUD_BASE_REPO}/${CODE_GEN_DIR}/push-cluster-state.sh"
     )