+++ Changelog.md
@@ -46,6 +46,7 @@
 - Add ArgoCD ApplicationSet support for multiple CDEs
 - Flag in env_var to enable/disable external ingresses for admin endpoints
 - Fix test that file has been copied is not working in init container
+- Update default version of pf-pingid-integration-kit to 2.24.0 in PingFederate
 - Deploy ArgoCD to customer-hub only
 - Fix the ArgoCD App name
 - Add ArgoCD IRSA role
@@ -58,6 +59,8 @@
 - Add priorityClassName into CAWAgent daemonset
 - Fluent-bit pods stuck in pending state
 - Add logstash and fluent-bit alerts
+- Fix Kibana Visualization "Ping Access - Response Codes Over Time"
+- No longer move files into custom-resources directory when upgrading
 
 _Changes:_
 
@@ -66,7 +69,9 @@ _Changes:_
 - [X] PDO-3834 Updated StorageClass type to gp3
 - [X] PDO-3908 Clean up P1 artifacts in Admin environment during CI/CD teardown
 - [X] PDO-4009 Update k8s StorageClass provisoner to use CSI driver
+- [X] PDO-4309 Add integration test for PingDirectory Java args 
 - [X] PDO-4388 Flag in env_var to enable/disable external ingresses for admin endpoints
+- [X] PDO-4548 Find and destroy file moving to custom-resources code from generate-cluster-state.sh to fix reoccuring issue when upgrading
 - [X] PDO-4556 [need before EKS 1.25] batch/v1beta1 CronJob is deprecated in v1.21+, unavailable in v1.25+; use batch/v1 CronJob
 - [X] PDO-4575 Upgrade ArgoCD to v2.5.5
 - [X] PDO-4697 Update cluster tools to latest version: cluster-autoscaler v1.23.0
@@ -84,6 +89,7 @@ _Changes:_
 - [X] PDO-4773 Update generate-cluster-state script to create base and region values.yaml files
 - [X] PDO-4774 Update generate-cluster-state script to massage the new code-gen structure files into the new CSR structure
 - [X] PDO-4775 Add new ArgoCD application definition to PCB
+- [X] PDO-4780 Move tag-release.sh and PCB ci-scripts to shared location
 - [X] PDO-4817 Remove SIEM console logging for PA/PF
 - [X] PDO-4818 Add multiple USER_BASE_DNs and BACKEND_IDs env vars
 - [X] PDO-4822 Add multiple backends to 'BACKENDS_TO_BACKUP' env var
@@ -112,6 +118,7 @@ _Changes:_
 - [X] PDO-4997 Update values.yaml in region and base path sync with env_vars file
 - [X] PDO-5008 Update tagging script to return correct tag
 - [X] PDO-5009 Add logstash HPA and upgrade resources
+- [X] PDO-5017 Use SUPPORTED_ENVIRONMENT_TYPES for generate/update scripts
 - [X] PDO-5018 PGO resources - handle secondary region v1.18
 - [X] PDO-5025 Improve fluent-bit multiline log parsing
 - [X] PDO-5037 Update to replace deprecated topologyKey to topology.kubernetes.io/zone 
@@ -119,6 +126,7 @@ _Changes:_
 - [X] PDO-5041 node-role.kubernetes.io/master (deprecated)
 - [X] PDO-5043 Legacy Logging Mode: Remove Feature Flag, Code Logic and Refactor Filters
 - [X] PDO-5080 Test that file has been copied is not working in init container
+- [X] PDO-5090 Update default version of pf-pingid-integration-kit to 2.24.0 in PingFederate
 - [X] PDO-5104 Update Prometheus CPU/MEM limits
 - [X] PDO-5107 Fluent-bit pods stuck in pending state
 - [X] PDO-5124 Enable bootstrapping a new customer with ArgoCD
@@ -128,6 +136,10 @@ _Changes:_
 - [X] PDO-5144 Add logstash/fluent-bit readiness/liveness probe
 - [X] PDO-5147 Add logstash metrics to prometheus
 - [X] PDO-5148 Modify Prometheus query for all backup alerting to only include the primary pod
+- [X] PDO-5191 Update image_map to align with tagging process
+- [X] PDO-5221 'Field "responseCode.keyword" not found' on the 'Ping Access - Response Codes Over Time' visualization
+- [X] PDO-5223 Remove pa-was config for ArgoCD from non customer-hub CDEs
+- [X] PDO-5261 Remove PF and PA from pa-was config in customer-hub
 
 ### 1.17.0.0